updated to output inference for more analysi

+++ features/datasets.py
@@ -204,7 +204,7 @@ class ImageDatasetInfer(Dataset):
             except:
                 print("ERRORORORORO")
                 print(self.smiles[item])
-                image = torch.from_numpy(np.zeros((3,128, 128)))
+                image = torch.from_numpy(np.zeros((3,128, 128))).float()
             # property = self.property_func(mol)
             #
             # # TODO align property
